+++ lib/src/event/event_bus.dart
@@ -16,11 +16,6 @@ part of dart_web_toolkit_event;
  */
 abstract class EventBus<H> {
 
-//  /**
-//   * @deprecated required by legacy features in GWT's old HandlerManager
-//   */
-//  bool isEventHandled(EventType<H> type);
-  
   /**
    * Invokes [dispatch] with [handler].
    *